market/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.2.7
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.0.6
django-admin-rangefilter==0.8.5
django-daterange-filter==1.3.0
django-model-utils==4.0.0
frozenlist==1.3.1
html5lib==1.0.1
idna==3.4
multidict==6.0.2
Pillow==7.1.2
psycopg2-binary==2.8.5
PyPDF==1.26.0
pyTelegramBotAPI==4.7.1
pytz==2020.1
reportlab==3.5.42
requests==2.28.1
six==1.14.0
sqlparse==0.3.1
telepot==12.7
typing-extensions==4.4.0
Unipath==1.1
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.4
yarl==1.8.1